# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for BrandishBot game engine - inventory, crafting, economy, and stats management",
        "title": "BrandishBot API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/osse101/BrandishBot_Go"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/user/inventory": {
            "get": {
                "description": "Get all items in a user's inventory",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "Get user inventory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetInventoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/item/buy": {
            "post": {
                "description": "Buy an item with currency",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "economy"
                ],
                "summary": "Buy item",
                "parameters": [
                    {
                        "description": "Buy details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.BuyItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.BuyItemResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Feature locked",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/item/give": {
            "post": {
                "description": "Transfer an item from one user to another",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "Give item to another user",
                "parameters": [
                    {
                        "description": "Transfer details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.GiveItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/item/sell": {
            "post": {
                "description": "Sell an item for currency",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "economy"
                ],
                "summary": "Sell item",
                "parameters": [
                    {
                        "description": "Sell details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SellItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SellItemResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Feature locked",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/item/use": {
            "post": {
                "description": "Use an item from inventory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory"
                ],
                "summary": "Use item",
                "parameters": [
                    {
                        "description": "Usage details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UseItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.UseItemResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.BuyItemRequest": {
            "type": "object",
            "required": [
                "item_name",
                "username"
            ],
            "properties": {
                "item_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "platform": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "maximum": 10000,
                    "minimum": 1
                },
                "username": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "handler.BuyItemResponse": {
            "type": "object",
            "properties": {
                "items_bought": {
                    "type": "integer"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handler.GetInventoryResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.InventoryItem"
                    }
                }
            }
        },
        "handler.GiveItemRequest": {
            "type": "object",
            "required": [
                "item_name",
                "owner",
                "receiver"
            ],
            "properties": {
                "item_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "owner": {
                    "type": "string",
                    "maxLength": 100
                },
                "platform": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "maximum": 10000,
                    "minimum": 1
                },
                "receiver": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
...
        "handler.SellItemRequest": {
            "type": "object",
            "required": [
                "item_name",
                "username"
            ],
            "properties": {
                "item_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "platform": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "maximum": 10000,
                    "minimum": 1
                },
                "username": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "handler.SellItemResponse": {
            "type": "object",
            "properties": {
                "items_sold": {
                    "type": "integer"
                },
                "money_gained": {
                    "type": "integer"
                }
            }
        },
        "handler.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.UseItemRequest": {
            "type": "object",
            "required": [
                "item_name",
                "username"
            ],
            "properties": {
                "item_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "platform": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "maximum": 10000,
                    "minimum": 1
                },
                "target_username": {
                    "type": "string",
                    "maxLength": 100
                },
                "username": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "handler.UseItemResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "user.InventoryItem": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "value": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}